% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookdown.R
\name{gitbook}
\alias{gitbook}
\title{rebuild your favorite book with custom tuning}
\usage{
gitbook(url, custom_format = NULL, repo = NULL)
}
\arguments{
\item{url}{string. url of the original book}

\item{custom_format}{string. path to the custom \code{_output.yaml} file}

\item{repo}{string. the}
}
