% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gitbook.R
\name{make_gitbook}
\alias{make_gitbook}
\title{rebuild your favorite book with custom tuning}
\usage{
make_gitbook(
  output_format_file = NULL,
  url = "https://example.com",
  download_link = "https://example.com"
)
}
\arguments{
\item{output_format_file}{string. path to the custom \verb{_output.yaml} file.}

\item{url}{string. url of the original book.}

\item{download_link}{string. the url of the compressed file of latest book.}
}
\value{
list. Customized \code{\link[bookdown:gitbook]{bookdown::gitbook()}}, a R Markdown output format.
}
\description{
rebuild your favorite book with custom tuning
}
\examples{
NULL
}
