% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gitbook.R
\name{make_gitbook}
\alias{make_gitbook}
\title{rebuild your favorite book with custom tuning}
\usage{
make_gitbook(url, output_format_file = NULL, download_link = NULL)
}
\arguments{
\item{url}{string. url of the original book}

\item{output_format_file}{string. path to the custom \code{_output.yaml} file}

\item{download_link}{string. the url of the compressed file of latest book. if \code{NULL}, will use zhuoer's GitLab, repository name deduced from \code{url}}
}
\description{
rebuild your favorite book with custom tuning
}
